requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
paho-mqtt==2.0.0
pillow==10.2.0
python-decouple==3.8
requests==2.31.0
urllib3==2.2.1
asyncio-mqtt==0.16.2
aiohttp==3.9.3
